test: share schtasks gateway script fixture
