+++ src/daemon/schtasks.startup-fallback.test.ts
@@ -10,6 +10,7 @@ import {
   resetSchtasksBaseMocks,
   schtasksResponses,
   withWindowsEnv,
+  writeGatewayScript,
 } from "./test-helpers/schtasks-fixtures.js";
 const childUnref = vi.hoisted(() => vi.fn());
 const spawn = vi.hoisted(() => vi.fn(() => ({ unref: childUnref })));
@@ -43,20 +44,6 @@ function resolveStartupEntryPath(env: Record<string, string>) {
   );
 }
 
-async function writeGatewayScript(env: Record<string, string>, port = 18789) {
-  const scriptPath = resolveTaskScriptPath(env);
-  await fs.mkdir(path.dirname(scriptPath), { recursive: true });
-  await fs.writeFile(
-    scriptPath,
-    [
-      "@echo off",
-      `set "OPENCLAW_GATEWAY_PORT=${port}"`,
-      `"C:\\Program Files\\nodejs\\node.exe" "C:\\Users\\steipete\\AppData\\Roaming\\npm\\node_modules\\openclaw\\dist\\index.js" gateway --port ${port}`,
-      "",
-    ].join("\r\n"),
-    "utf8",
-  );
-}
 async function writeStartupFallbackEntry(env: Record<string, string>) {
   const startupEntryPath = resolveStartupEntryPath(env);
   await fs.mkdir(path.dirname(startupEntryPath), { recursive: true });
